Struct gdkx11::builders::X11DeviceXI2Builder

pub struct X11DeviceXI2Builder { /* private fields */ }

A builder-pattern type to construct X11DeviceXI2 objects.

Implementations

impl X11DeviceXI2Builder

pub fn device_id(self, device_id: i32) -> Self

pub fn device_manager(self, device_manager: &impl IsA<DeviceManager>) -> Self

The gdk::DeviceManager the gdk::Device pertains to.

pub fn display(self, display: &impl IsA<Display>) -> Self

The gdk::Display the gdk::Device pertains to.

pub fn has_cursor(self, has_cursor: bool) -> Self

Whether the device is represented by a cursor on the screen. Devices of type gdk::DeviceType::Master will have true here.

pub fn input_mode(self, input_mode: InputMode) -> Self

pub fn input_source(self, input_source: InputSource) -> Self

Source type for the device.

pub fn name(self, name: impl Into<GString>) -> Self

The device name.

pub fn num_touches(self, num_touches: u32) -> Self

The maximal number of concurrent touches on a touch device. Will be 0 if the device is not a touch device or if the number of touches is unknown.

pub fn product_id(self, product_id: impl Into<GString>) -> Self

Product ID of this device, see [DeviceExtManual::product_id()][crate::gdk::prelude::DeviceExtManual::product_id()].

pub fn seat(self, seat: &Seat) -> Self

gdk::Seat of this device.

pub fn type_(self, type_: DeviceType) -> Self

pub fn vendor_id(self, vendor_id: impl Into<GString>) -> Self

Vendor ID of this device, see [DeviceExtManual::vendor_id()][crate::gdk::prelude::DeviceExtManual::vendor_id()].

pub fn build(self) -> X11DeviceXI2

Build the X11DeviceXI2.