gdkx11
Function x11_device_get_id

pub fn x11_device_get_id(device: &X11DeviceCore) -> i32

Returns the device ID as seen by XInput2.

If gdk_disable_multidevice() has been called, this function will respectively return 2/3 for the core pointer and keyboard, (matching the IDs for the Virtual Core Pointer and Keyboard in XInput 2), but calling this function on any slave devices (i.e. those managed via XInput 1.x), will return 0.

device

a gdk::Device

Returns

the XInput2 device ID.