gdkx11

Function x11_register_standard_event_type

pub fn x11_register_standard_event_type(
    display: &X11Display,
    event_base: i32,
    n_events: i32,
)

Registers interest in receiving extension events with type codes between event_base and event_base + n_events - 1. The registered events must have the window field in the same place as core X events (this is not the case for e.g. XKB extension events).

If an event type is registered, events of this type will go through global and window-specific filters (see [WindowExtManual::add_filter()][crate::gdk::prelude::WindowExtManual::add_filter()]). Unregistered events will only go through global filters. GDK may register the events of some X extensions on its own.

This function should only be needed in unusual circumstances, e.g. when filtering XInput extension events on the root window.

display

a gdk::Display

event_base

first event type code to register

n_events

number of event type codes to register