Crate gdkx11

Rust GDKX11 bindings

This project is UNMAINTAINED. Please take a look at gtk4-rs instead!

Project site is here.

Rust bindings and wrappers for GDKX11, part of gtk3-rs.

GDKX11 3.22 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.70.0.

Documentation

Rust API - Stable

Rust API - Development

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gdkx11 = { git = "https://github.com/gtk-rs/gtk3-rs.git", package = "gdkx11" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
gdkx11 = "0.13"
gdkx11 = { git = "https://github.com/gtk-rs/gtk3-rs.git", package = "gdkx11" }

See Also

glib

gio

gdk

License

gdkx11 is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use gdk;

pub use gio;

pub use glib;

pub use x11;

Modules

builders

Builder pattern types.

functions

Structs

X11AppLaunchContext

X11Cursor

X11DeviceCore

X11DeviceManagerCore

X11DeviceManagerXI2

X11DeviceXI2

X11Display

X11DisplayManager

X11DragContext

X11GLContext

X11Keymap

X11Monitor

X11Screen

X11Visual

X11Window