Crate gdkx11

Rust GDKX11 bindings

Project site is here.

Rust bindings and wrappers for GDKX11, part of gtk3-rs.

GDKX11 3.22 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.70.0.

Documentation

Rust API - Stable

Rust API - Development

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gdkx11 = { git = "https://github.com/gtk-rs/gtk3-rs.git", package = "gdkx11" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
gdkx11 = "0.13"
gdkx11 = { git = "https://github.com/gtk-rs/gtk3-rs.git", package = "gdkx11" }

See Also

glib

gio

gdk

License

gdkx11 is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use gdk;

pub use gio;

pub use glib;

pub use x11;

Modules

builders

Builder pattern types.

functions

Structs

X11AppLaunchContext

Implements

X11Cursor

Implements

X11DeviceCore

Implements

X11DeviceManagerCore

Implements

X11DeviceManagerXI2

Properties

X11DeviceXI2

Properties

X11Display

Implements

X11DisplayManager

Implements

X11DragContext

Implements

X11GLContext

Implements

X11Keymap

Implements

X11Monitor

Implements

X11Screen

Signals

X11Visual

Implements

X11Window

Implements