Struct gdkx11::X11Window

#[repr(transparent)]
pub struct X11Window { /* private fields */ }

Implements

[trait@gdk::prelude::WindowExt]

Implementations

impl X11Window

pub fn foreign_new_for_display( display: &X11Display, window: Window ) -> X11Window

Wraps a native window in a gdk::Window. The function will try to look up the window using lookup_for_display() first. If it does not find it there, it will create a new window.

This may fail if the window has been destroyed. If the window was already known to GDK, a new reference to the existing gdk::Window is returned.

display

the gdk::Display where the window handle comes from.

window

an Xlib Window

Returns

a gdk::Window wrapper for the native window, or None if the window has been destroyed. The wrapper will be newly created, if one doesn’t exist already.

pub fn desktop(&self) -> u32

Gets the number of the workspace self is on.

Returns

the current workspace of self

pub fn xid(&self) -> Window

Returns the X resource (window) belonging to a gdk::Window.

Returns

the ID of drawable’s X resource.

pub fn move_to_current_desktop(&self)

Moves the window to the correct workspace when running under a window manager that supports multiple workspaces, as described in the Extended Window Manager Hints specification. Will not do anything if the window is already on all workspaces.

pub fn move_to_desktop(&self, desktop: u32)

Moves the window to the given workspace when running unde a window manager that supports multiple workspaces, as described in the Extended Window Manager Hints specification.

desktop

the number of the workspace to move the window to

pub fn set_frame_sync_enabled(&self, frame_sync_enabled: bool)

This function can be used to disable frame synchronization for a window. Normally frame synchronziation will be enabled or disabled based on whether the system has a compositor that supports frame synchronization, but if the window is not directly managed by the window manager, then frame synchronziation may need to be disabled. This is the case for a window embedded via the XEMBED protocol.

frame_sync_enabled

whether frame-synchronization should be enabled

pub fn set_hide_titlebar_when_maximized( &self, hide_titlebar_when_maximized: bool )

Set a hint for the window manager, requesting that the titlebar should be hidden when the window is maximized.

Note that this property is automatically updated by GTK+, so this function should only be used by applications which do not use GTK+ to create toplevel windows.

hide_titlebar_when_maximized

whether to hide the titlebar when maximized

pub fn set_theme_variant(&self, variant: &str)

GTK+ applications can request a dark theme variant. In order to make other applications - namely window managers using GTK+ for themeing - aware of this choice, GTK+ uses this function to export the requested theme variant as _GTK_THEME_VARIANT property on toplevel windows.

Note that this property is automatically updated by GTK+, so this function should only be used by applications which do not use GTK+ to create toplevel windows.

variant

the theme variant to export

pub fn set_user_time(&self, timestamp: u32)

The application can use this call to update the _NET_WM_USER_TIME property on a toplevel window. This property stores an Xserver time which represents the time of the last user input event received for this window. This property may be used by the window manager to alter the focus, stacking, and/or placement behavior of windows when they are mapped depending on whether the new window was created by a user action or is a “pop-up” window activated by a timer or some other event.

Note that this property is automatically updated by GDK, so this function should only be used by applications which handle input events bypassing GDK.

timestamp

An XServer timestamp to which the property should be set

pub fn set_utf8_property(&self, name: &str, value: Option<&str>)

This function modifies or removes an arbitrary X11 window property of type UTF8_STRING. If the given self is not a toplevel window, it is ignored.

name

Property name, will be interned as an X atom

value

Property value, or None to delete

pub fn lookup_for_display( display: &X11Display, window: Window ) -> Option<X11Window>

Looks up the gdk::Window that wraps the given native window handle.

display

the gdk::Display corresponding to the window handle

window

an Xlib Window

Returns

the gdk::Window wrapper for the native window, or None if there is none.